requirements.txt
fastapi
uvicorn
pydantic
pytest-html
sqlalchemy
uvicorn
coverage
redis
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary 